Actions organized into separate files. Unused actions removed.
